backendd/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

python-multipart==0.0.19
